Fixes issue #2249, ProviderRegistryClass to separate file
Moves the ProviderRegistry class into its own file.  This does
not change any functionality or tests.
